llm/server/whisper_server/requirements.txt
fastapi
uvicorn
click
openai-whisper
python-multipart